pyproject.toml
[tool.poetry]
name = "et_fpm"
version = "0.1.0"
description = "Energy-Transition Financial Performance Monitor"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "et_fpm"}]

[tool.poetry.dependencies]
python = "^3.11"
yfinance = ">=0.2"
pandas = ">=2.2"
numpy = ">=1.26"
sqlalchemy = ">=2.0"
psycopg2-binary = "*"
timescale-python = "*"
prefect = ">=2.18"
statsmodels = ">=0.14"
pyyaml = "*"
scipy = "*"
streamlit = ">=1.34"
plotly = ">=5.20"
weasyprint = "*"
python-dotenv = "*"
pydantic = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
black = "*"
ruff = "*"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
